remove remote-geolocation-client and move back geolocation handling into remote-simulator-client
